Recognize 1904 and 1900 as potential base dates in xlsx
Modern excels use 1900 as the base date when storing dates as numbers,
but Excel for Mac < 2011 uses 1904! Roo already supported the
different offsets in .xls files but didn't support the offsets in
.xlsx format. This change adds support for recognizing the base date
format in .xlsx files.

1904 vs 1900 is configured in the workbookPr node of workbook.xml.

    http://msdn.microsoft.com/en-us/library/ff530155(v=office.12).aspx

There will be no difference in behavior for any spreadsheet generated
by any version of Windows excel or Excel for Mac 2011.